Rename timeline.tsx to timeMarker.tsx
